feat(zudo-doc): ship color-scheme + design-token-panel mechanism to package (#2333)
Move the palette→--zd-* CSS injection mechanism and zdtp wiring from the
project to @takazudo/zudo-doc, keeping all DATA project-side.

New package exports:
- @takazudo/zudo-doc/color-scheme-utils — ColorRef, ColorScheme types,
  SEMANTIC_DEFAULTS, SEMANTIC_CSS_NAMES, resolveColor, resolveSemanticColors,
  schemeToCssPairs, generateCssCustomProperties(scheme), generateLightDarkCssProperties(light,dark)
- @takazudo/zudo-doc/design-token-panel-bootstrap — bootstrapDesignTokenPanel(panelConfig)
  that wires configurePanel + setLifecycleAdapter to zfb navigation events

Host src/config/color-scheme-utils.ts becomes a thin wrapper importing the
mechanism from the package and adding settings-bound convenience functions.
src/lib/design-token-panel-bootstrap.ts delegates to bootstrapDesignTokenPanel().

Templates updated in sync (base color-scheme-utils, designTokenPanel feature
bootstrap). Template drift check and build green. --zd-* injection values
unchanged (confirmed: light-dark(#e2ddda, #181818) for --zd-bg).
